Generate id on client when create object, Fix processing 64-bit numbers in json

+++ platform/shared/json/json_tokener.c
@@ -396,9 +396,9 @@ struct json_object* json_tokener_parse_ex(struct json_tokener *tok,
 	printbuf_memappend(tok->pb, &c, 1);	
 	if(c == '.' || c == 'e' || c == 'E') tok->is_double = 1;
       } else {
-	int numi;
+	int64 numi;
 	double numd;
-	if(!tok->is_double && sscanf(tok->pb->buf, "%d", &numi) == 1) {
+	if(!tok->is_double && sscanf(tok->pb->buf, "%lli", &numi) == 1) {
 	  current = json_object_new_int(numi);
 	} else if(tok->is_double && sscanf(tok->pb->buf, "%lf", &numd) == 1) {
 	  current = json_object_new_double(numd);
